requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1

requests==2.32.3
psycopg2-binary==2.9.9
pdfplumber==0.11.0
python-multipart==0.0.9

pydantic==2.8.2
pydantic-settings==2.4.0
